remove duplicate zeebu
